Revamped Task Management System
Tasks with NPCs and Locations are much more customizable with many less restrictions.
- Task NPCs and Locations no longer tied to enum. Meaning adding new subjects and locations becomes more streamlined
- Task Subjects now tied to task Types, meaning strange productions like "Buy weather from the Baker" are no longer possible.
For each subject dictionary, tasks types are added to provide context for tasks. So, in the future:
For the Baker, given the type of "Buy" we can be assigned to buy "Bread" or "Pastry" or "Cake"
For the same Baker, given the type of "Ask" we can be assigned to ask about the Bakers favorite "Color" or "Sport" or "Car"